[RIT-115] Refactor architecture for better enrichment (#310)
* [RIT 115] Refactor db architecture

* [RIT 115] Enrichment V1

+++ apps/api/src/routes_web/lists/get_all.ts
@@ -15,7 +15,7 @@ export const getLists = async (
         id: list.id,
         name: list.name,
         emoji: list.emoji,
-        itemCount: sql<number>`count(${listPlace.id})::int`,
+        itemCount: sql<number>`count(${listPlace.id})`,
         createdAt: list.createdAt,
         updatedAt: list.updatedAt,
       })
